[surface model] Enable safety margin for bounding boxes in BVH
Using single precision bounding boxes, a safety margin helps preventing errors.

In fact, we found an error that appears only in release mode (due to compiler optimization that leads to slightly different results, such that the ray was inside the bounding box, causing errors).

This ray with the following geometry fails in release mode but works in debug mode: [CMS_simple24.gdml](/uploads/d0971dd07fafcb75bdbd227e95a8636a/CMS_simple24.gdml)
```
./testRaytracing -gdml_name ../../gdml/CMS_simple24.gdml -nrays 1000000 -ongpu 0 -detect_overlaps 1 -mmunit 0.1 -verbosity 1 -debug 1  -accept_zeros 1  -test_bvh 1 -direction -0.9891568946458417, -0.1465515971014428, -0.009553384824854797 -point 273.71362420435526 -98.71404743857443 564.2559375072851  -max_cross 1
```

* enable safety margin for bounding boxes in BVH

[surface][mini] fix a few pedantic compiler warnings

Sometimes it is helpful to compile with pedantic compiler warnings for debugging.

This MR fixes a few warnings with pedantic compiler warnings set on. Note that are still a ton of those but this fixes some of the most obvious ones that are printed very often.

* remove redundant semicolons