[surface model] Fix crucial memory error. Heap on GPU increased, fix 0 allocations, better errors. BVH validation.
In the `IsConvexConcave` there was a critical bug to be observed using this geometry: [CMS_polyhedron3.gdml](/uploads/355d9538d68636f297129866237df019/CMS_polyhedron3.gdml)

In the first segment, the index could end up being -1 due to a missing check, leading to memory corruption.

In this example, the corrupted memory was the bounding box of the surface BVH, so in the memory investigation I added a few improvements around that region:
1.
```
    std::vector<LogicalVolume const *> logicalvolumes(container.size());
    logicalvolumes.resize(0);
    for (auto p : container) {
      logicalvolumes.push_back(p.second);
    }
```
The allocation of container.size() was overwritten by the resize(0). The container.size() reservation was redundant as it was removed by the resize(0) right away.
The best way is to reserve the memory (which doesn't initialize with 0 pointers) and then pushback.
2. the shell was copied when accessed, now it is accessed via reference and made const to avoid unwanted changes.

* fix crucial memory error

increase heap on GPU, add error messages, fix 0 allocations on GPU

The CMS2026D110 geometry fails with the solid model on GPU.

The reason is that the allocation in the PlanarPolygon for the simple extruded exceeds the available heap (which is at the default limit of 8 MB) on the GPU. Therefore, the limit must be increased.
Additionally, there were allocations of 0 size in the PlanarPolygon, causing memory errors.

This MR adds a few asserts, so that the reason for the crash can be detected. Note that the detection only works when run with `compute-sanitizer` and is not even caught by running just in debug mode.

Note: This MR should be cherry-picked into the master. Also, quite a few changes are only due to clang formatting.

* use power of 2

* cleaning

* remove unused variable, put reserve and resize for vectorization back in place

* increase heap on GPU, add error messages, fix 0 allocations on GPU

[surface model] improve error description and handling on GPU, do validation of BVH also on GPU

This MR improves the error handling on GPU. It adds the option `-max_cross` also on GPU and adds a validation of solid model vs surface model with BVH on GPU.
Both errors are then printed separately, so one can see which navigator has how many errors.
Also, an primitive accept_zero option was added for the gpu error calculation. It simply doesn't count the 0 steps to the num_cross. Note that a 0 step is defined as 1e-6 (1000 * kTolerance). This can still lead to errors, if the solid model does a step of 1.005e-6 and the surface model 0.995e-6 (which was observed before). This accept_zero option is not as advanced as the one on host but it does a decent job.

* add accept_zero option for GPU

* improve error description and handling on GPU, do validation of BVH also on GPU