increase stack for cms gpu CI test
When running `ctest` in debug mode, the test `GeometryTest-Device-cms2018` fails. In fact, it is probably also erroneous in release mode but it goes undetected.
With `compute-sanitizer` it was revealed that in fact the stack overflows, therefore increasing the stack fixes the issue. There is a typo in the commit message, it should of course be stack not heap.

* use power of 2

* increase heap for cms gpu CI test