[surface model] Do surface generation always in vecgeom::Precision
Previously, the common surface generation and calculating of the masks and surface data was done in single precision in mixed precision mode.

That lead to vastly different settings of common surfaces in single and double precision, e.g, for the cms2026D110 gdml we got:
```
SINGLE PRECISION
Surface model info:  13133900 touchables, 123 scenes
    volume shells          = 13934 [1.30896 MB]
    local transformations  = 109059 [10.8167 MB]
    global transformations = 1603616 [159.05 MB]
    local surfaces         = 120193 [5.9605 MB]
    global surfaces        = 1603615 [79.525 MB]
    common surfaces        = 272147 [24.8042 MB]
    candidates             = 4974134 [18.9748 MB]
    window masks           = 405515 [6.18767 MB]
    cyl/sph masks          = 11838 [0.587059 MB]
    ring masks             = 12064 [0.322144 MB]
    Z/phi masks            = 23265 [0.79874 MB]
    triangle masks         = 13108 [0.600037 MB]
    quad masks             = 17692 [1.07983 MB]
    BVHs                   = 13934 [39.5015 MB]
 Total: 349.517[MB]
__________________________

DOUBLE PRECISION
info: ___________________________________________________________________________________
 Surface model info:  13133900 touchables, 123 scenes
    volume shells          = 13934 [1.30896 MB]
    local transformations  = 109059 [10.8167 MB]
    global transformations = 1603616 [159.05 MB]
    local surfaces         = 120193 [5.9605 MB]
    global surfaces        = 1603615 [79.525 MB]
    common surfaces        = 309627 [27.3777 MB]
    candidates             = 5057907 [19.2944 MB]
    window masks           = 445126 [13.5842 MB]
    cyl/sph masks          = 11838 [0.587059 MB]
    ring masks             = 12064 [0.644287 MB]
    Z/phi masks            = 23297 [1.59968 MB]
    triangle masks         = 13108 [1.20007 MB]
    quad masks             = 17693 [2.15979 MB]
    BVHs                   = 13934 [39.5015 MB]
 Total: 362.61[MB]
```

This MR changes the fCPUdata to always be in the compiled `vecgeom::Precision`. Then, only in `UpdateSurfData`, when the fCPUdata is copied to the fSurfData, the masks, transformations and surface data containers are cast into single precision (if run in mixed precision mode). For this quite a few new constructors needed to be added. Note that with this MR, the fCPUdata can be switched back to `Real_t` with just changing the BrepHelper fCPUdata definition and instantiation.

With this MR, we now get the same amount of surfaces, masks etc as we get in double precision when we run in mixed precision (for cms2026D110):
```
SINGLE PRECISION
 Surface model info:  13133900 touchables, 123 scenes
    volume shells          = 13934 [1.30896 MB]
    local transformations  = 109059 [10.8167 MB]
    global transformations = 1603616 [159.05 MB]
    local surfaces         = 120193 [5.9605 MB]
    global surfaces        = 1603615 [79.525 MB]
    common surfaces        = 309627 [27.3777 MB]
    candidates             = 5057907 [19.2944 MB]
    window masks           = 445126 [6.79208 MB]
    cyl/sph masks          = 11838 [0.587059 MB]
    ring masks             = 12064 [0.322144 MB]
    Z/phi masks            = 23297 [0.799839 MB]
    triangle masks         = 13108 [0.600037 MB]
    quad masks             = 17693 [1.0799 MB]
    BVHs                   = 13934 [39.5015 MB]
 Total: 353.016[MB]
```

This fixes quite a few errors in many solids when running in mixed precision mode.

The correctness of this MR was further tested by running CMS with 10 mio rays (in full double precision) with overlap detection and having 0 errors (as before).

* do surface generation always in vecgeom::Precision