[surface model] Use relative push for booleans
When running in mixed precision, more fine-tuning of the tolerances is needed in the mask checks, distance checks etc.

Furthermore, the boolean push has proven quite problematic in mixed precision, since the single precision tolerance is very high, leading to a large push. After finding failing rays for a too large and for a too small push, here I propose to use a relative push to the distance. So far, this seems to be more stable, but still a huge amount of rays needs to be debugged, so this is work in progress and will probably require more refinement.

* use relative push for booleans