[surface model] Store the volume transformation as separate to not access the solid model
Previously, the transformation to the placed volume in the BVH was taken directly from the solid model, which prevented it to be in single precision. Here, this is now separated (along with the logical volume id), so that the solid model doesn't need to be accessed in some places in the BVH.

In double precision, no performance difference was observed. In mixed precision, the simple examples showed a 15% speedup.

* forgot one file

* clearer naming

* fix namings

* fix naming

* store the volume transformation as separate to not access the solid model