[surface model] fix logic evaluation for known surfaces
Since we know that we were exiting / entering a surface, we can exclude that surface from the logic evaluation in booleans.
However, this didn't work properly at the moment, since logic ids can be negative (for flipped booleans) and thus the comparison was broken.
Also, it didn't work if the boolean was negated.

This MR fixes the issues.

Now, well behaved geometries can work without a push (LHCb, CMS test beam (not CMS itself)).

Despite that this MR is correct, it uncovers a rare problem with CMS that leads to stalling in the BVH, and therefore should not be merged for now

* fix logic evaluation for known surfaces