[surface model] Some BVH traversal improvements
Some simple improvements that reduce the amount of traversal steps for the BVH:

BVHsurf:
- `bstep` wasn't properly updated after finding a hit
- Nodes were visited in the wrong order, with the farthest node being added last to the traversal stack, and checked first
- Nodes are now checked when taken from the stack, if they are farther than the current closest hit, the branch is cut from traversal. Without this functionality, branches are checked until we reach the leaf nodes

BVH:
- For the solids BVH we already had the branch pruning functionality, and the step is updated correctly, the only improvement that applies is the change to the traversal order

When measured for `BVHsurf`, these changes reduce the total number of children visited during traversal:

| TrackML, 10000 rays/logical volume | Total visits |
| ------ | ------ |
| Current master | 2890717 |
| `bstep` update fix | 2574281 |
| Visit order + branch pruning | 2444753 |

In terms of performance however there are no visible changes for the surface traversal. For the solids traversal there seems to be a speedup of 1.08 on CPU, and 1.12 on GPU when applying the traversal order fix, when the aplication is compiled in debug mode, in release mode there doesn't seem to be an appreciable difference.

* Some BVH traversal improvements