Added ABBoxManager_t alias for forward compatibility with a templated version
In the `surface_model` branch, the singleton class `ABBoxManager` becomes templated on a precision type to support single precision in the surface BVH independently on the compile time `vecgeom::Precision`. This MR introduces `ABBoxManager_t`, an alias of the actual ABBoxManager version, to be used in dependent libraries for forward compatibility until the `surface_model` branch is merged.

So use: `vecgeom::ABBoxManager_t::Instance()` instead of `vecgeom::ABBoxManager::Instance()`

* Added ABBoxManager_t alias for forward compatibility with a templated version