[surface model] Added SideDivision helper structure for accelerating frame searches on CS sides.
This introduces division helpers at the level of CS sides, to accelerate the loops over frames surfaces during traversal. Some optimizing heuristics is used to select the best division axis. For the moment we support X, Y, R (when only ring frames are on one side), and XY divisions for planar surfaces, and Z divisions for cylindrical/conical surfaces.

* Small extra reshuffling of BrepHelper and CommonTypes, to clean-up the public interface for converting to the surface model.

* Implemented kXY division for planar surfaces.

* Implemented kZPhi division on Z (phi still missing)

* Made testEM3 example use BVH surface navigation

* Fixed downcasting to float error in BVH traversal.

* Latest fixes for the division helper

* Data structure for side divisions filled and copied to GPU

* Filling SlideDivisionCPU with optimal slicing of the extent.

* First implementation of planar side division

* Added SideDivision helper structure for accelerating frame searches on CS sides.

Small reshuffling of BrepHelper and CommonTypes, to clean-up the public interface for converting to the surface model.