[surface model] Use looser tolerance on approach in BVH
In some recent mixed precision improvements, the tolerance used in relative tolerance checks was set to be more strict.

However, when calculating the approach in single precision in the BVH, the approach should always be slightly reduced to avoid that the ray overshoots due to precision issues. This relative reduction should use the less strict tolerance of `kToleranceDist` over `kToleranceStrict`. This MR allows for an input parameter in those relative tolerance checks and sets it accordingly.

This fixes some rays in CMS, which caused a relocation in the BVH otherwise.

* use looser tolerance on approach