[surface model] Crop bounding boxes for BVH
This MR proposes to crop the bounding boxes of surfaces by the extent of the volume that the surface belongs to. That means that virtual boolean surfaces that exceed the size of the actual solid, are cropped or even completely cut away.

This reduces the number of boolean hits in raytracing and is a stepping stone to marking convex boolean surfaces as safe.

* crop bounding boxes for BVH