[surface model] make convex booleans normal surfaces
This MR introduces the new concept of convex boolean surfaces. If a boolean surface is convex, the boolean is always entered or exited if the surface is hit. Such surfaces do not need to trigger the boolean check, but instead can simply be treated as normal surfaces.

This MR also changes a few things in the BVH: the creation now relies on the CPUdata not on the surfdata anymore. 
The pruning of the BVH bounding boxes is split from the creation of them, since in some case the convexity needed to be determined before the pruning.
Also, the bounding boxes are not increased in the BVH anymore, instead in the approach a larger negative tolerance is applied in the approach.

This MR gives hardly any speedup in the big geometries since it is quite conservative in determining the convexity. However, in the creation, a significant speedup in LHCb was observed, while giving full correctness. Fixing CMS lost the speed, but I will work on recovering it again.

* fix clang formatting

* make convex booleans normal surfaces