[surface model] remove overlooked printouts, increase tolerance
This MR fixes the tolerance to be reasonable for the convexity check of booleans. Previously, corners could break convexity due to numerical reasons.

This improves the rate of convex boolean surfaces in LHCb from 16 to 25% and in CMS from 5 to 7%, leading to some (minor) speedups.

Furthermore, a few overlooked printouts are removed.

* remove overlooked printouts, increase tolerance