[surface model] add bounding boxes and improved intersection check
This MR improves the embedded check for frames. Before, it was failing for and returning intersections for segments that were clearly not intersecting. 
This MR reduces the warnings for non-embedded frames having embedding parents in CMS2026D110 from 226 to 168.

The run time is not affected (if at all, it is slightly faster now).

* add bounding boxes and improved intersection check