[surface model ] check embeddedness only after all parents have been checked
This MR fixes a few more false positive warnings about non-embedded surfaces in embedding parents.

The warning must only be printed after all parents have been checked, since a surface could be embedded in another parent. Furthermore, if there are multiple parents, we cannot print the warning, since on a polycone, a surface could have multiple parents and be embedded only in the union of them. Thus, we exclude surfaces with multiple parents for now.

The warning in the ring frame mask check was removed for the same reason as above, the warning is false positive if later a parent is found where the daughter is indeed embedded.

After this MR, only very few warnings are left for CMS, which are numerical overlaps.

* remove redundant line

* check embeddedness only after all parents have been checked, remove warning for multiple parents, remove warning for a single ring check