[surface model] BVH locate.
Adding the `LocatePointIn` call for the surface model, using `BVHsurf`

It is not possible to do the location using surface AABBs, so we add a secondary BVH built from solid AABBs. The algorithm changes compared to the one used for solids navigation, however, as it avoids using solid-model objects where possible.

The main changes are:

- Storing solid-based BVHs in SurfData, per Logical Volume
- Storing number of daughter PVs, global IDs, and transformations in volume shells
- New `LocatePointIn` call in `BVHSurfNavigator` and the necessary `ItemID` and `CandidateContains` callbacks

* Moving locate call under the test_bvh option

* Fixing last errors

* BVH Locate working with some errors

* Storing daughter PVol IDs in volume shells

* Initial changes for a single-precision solids BVH