[surface model] use points in global coordinate system always in double in mixed precision
When running in mixed precision, we found that the points in the global coordinate system to keep track of the ray must be in double, as otherwise large distances get very imprecise and produce errors.

With this, the test beam setup runs with 7 errors per 1 mio rays (and crashes for more, since more fixes are needed).

However, we found that in this setup, the speedup is already non-existing, which we assume is due to the fact that we are completely memory-bound for larger setups. Thus, the memory accesses need to improved next.

For small setups, mixed precision still gives a good speedup (up to factor of 2).

The following steps would be next for improving mixed precision:
1. use the point in double to push the approach.
2. rescale quadratic equations (cylinder, cone)
3. recalculate quadratic equations for small coefficients.

* cleaning

* revert to full double precision

* fix merge conflict leftovers

* use points in global coordinate system always in double in mixed precision