[surface model] Add single step mode for surface model, including a split kernel approach
To improve profiling and make it for a more realistic case, this MR introduces a new mode in the testRaytracing. Using `-bvh_single_step 1`, the propagation is done with one kernel launch per step, as it is done in AdePT for a real simulation. With verbosity > 3, the timings and number of alive particles per step are printed.

Note that this single step approach has a larger overhead due to two kernel launches, so for simulations with long tails (were many steps with very few particles are done) this approach is slower. For a balanced case, the single step approach is faster.

Additionally, the same approach is done with a split kernel for the transportation with the BVH. The first kernel finds the framed surface, the second kernel does the relocation. The idea was to reduce register usage and thereby improve occupancy, but the register usage for the split approach is not sufficient to hit another breakpoint (64 or 128 registers per thread) and therefore the occupancy is unchanged, while the overhead increases the run time.

Next steps could be to also add a single step approach for the solid model with BVH and do add validation of the rays.


Small note: an unused parameter in the BVH was commented out.

* fix default value

* fix usage of long

* fix clang formatting

* adapt to new vol index system

* Add single step mode with BVH, add single step mode with kernel split