[surface model] add safeguard for parallel rays in BVH
In the BVH, parallel rays hitting the corner of multiple boxes could lead to infinite loops:
![image](/uploads/80b4a0f96894fc549d98156d19738add/image.png)
If a ray coming from the bottom was hitting exactly the corner, it would hit one of the boxes, then do a 0 step to exit through one of the side boxes, thereby be relocated to the other box. Since with a 0 step it could enter multiple boxes, this lead to an infinite loop.

Also, if a ray came with an angle almost perpendicular, the relative tolerance added to the BVH bounding boxes was not enough to capture it, then no exiting surface would be found, the ray would be relocated and then could enter with another 0 step.

Those two problems are fixed by a) increasing the tolerance for bounding boxes to see if there is an intersection and b) by adding a safeguard that parallel rays with a 0 step are discarded.

* improve comment

* increase tolerance for bounding boxes and add safeguard for parallel rays