[surface model] fix navigation in looper
There was a bug in the navigation of the looper. It was unnoticed, since the overlap detection would fix it.

When relocating to new scenes, the onsurf must be computed. This was done using the TopMatrix. However, case of scenes, the point that was passed was already in the scene coordinates, leading to incorrect final results. 

This MR fixes this. Now, the onsurf is calculated correctly from the global coordinates.

Furthermore, boolean volumes are forbidden to become scene volumes (this caused errors).

A small performance improvement was added: in the BVH, one doesn't need to call the TopMatrix, but instead can take the scene matrix and just apply additionally the TopInSceneMatrix.

* cleaning

* fix navigation in looper