[surface model] Splitting BVHSurfNavigator::ComputeStepAndHit in finding the distance and relocation methods
AdePT navigation needs splitting the ComputeStepAndHit method in a part returning the distance and setting the correct on_boundary flag, and another doing the actual relocation and computing the output state. This MR splits this functionality in ComputeStepAndNextSurface, returning the distance and the index of the hit surface for the current volume, and RelocateToNextVolume, using these values to relocate.

* fix merge conflicts

* [surface model]

The last exited state should be set only if we actually hit a surface and not if we leave the function earlier due to the physics limit

* remove asserts again

* fix setting of lastexited and use asserts

* Removed input state from Relocate interface

* Splitting BVHSurfNavigator::ComputeStepAndHit in finding the distance and relocation methods

AdePT navigation needs splitting the ComputeStepAndHit method in a part returning the distance and setting the correct on_boundary flag, and another doing the actual relocation and computing the output state. This MR splits this functionality in ComputeStepAndNextSurface, returning the distance and the index of the hit surface for the current volume, and RelocateToNextVolume, using these values to relocate.