[surface model] remove conditional in BVH approach
Currently, the approach is calculated like this
```
    approach = (hit && tmin > vecgeom::kToleranceDist<Real_t>)
                   ? vecgeom::MakeMinusTolerantRel(tmin, 10 * vecgeom::kToleranceDist<Real_t>)
                   : Real_t(0);
```
However, the BVH approach could be negative, since `vecgeom::MakeMinusTolerantRel(tmin, 10 * vecgeom::kToleranceDist<Real_t>)` potentially subtracts more than `tmin > vecgeom::kToleranceDist<Real_t>`. Since the approach is only used if `hit = true`, we can simplify this and remove the conditional entirely and assign the approach to the Max of 0 and how the approach was calculated previously.

This slightly improves performance as well.

* get rid of redundant if

* remove conditionals