[surface model] Use 2D transformations for relocation
In the relocation, we first do a transformation to the common surface. Then, when doing the frame checks, we do 3D transformations into the frame. However, since the frame sits on the common surface, only 2D transformations are needed. This MR replaces the 3D transformation on the side of the CS with 2D transformations. This speeds up the code when traversing CS with many frames. It slightly speeds up the TB setup but does not seem to impact LHCb or CMS, probably because the slowest surface on the warp is a boolean or something else and doesn't benefit from this effect.

The correctness was tested with CMS, LHCb and TB.

* remove cross precision transformations

* mini cleaning

* cleaning

* first implementation

* just templating framed surface and adding 2D transformations, not using them