Add intelligent cmake defaults and remove "surprising" ccache usage
This adds a new `vecgeom_set_default` cmake command which will:
- set a value for the given variable if undefined
- allow user cache variables to override it
- save a descriptive OPTION in the cmake cache if VecGeom is the top-level project
- output the defaults that it's setting

The new defaults for RPATH variables will fix when a shared-lib install is used downstream.

It also removes a default of secretly looking and using ccache, which is not appropriate outside of a development environment (and can easily be set by the user).

* Add intelligent cmake defaults