[surface model] Divisions are now computed based on CPUdata
The divisions used for the common surface sides were computed on the final SurfData structure. This is bad because the final frames depend on the precision types, and their transformations are already 2D. This PR moves the computation in the surface preparatory phase, using the temporary structure CPUdata. This is needed for pre-computing additional traversal helpers.

* [surf_model] Divisions are now computed based on CPUdata