[surface model] Framed surface transition precalculation
This implements the pre-calculation of **non-Boolean** framed surface traversals on **non-scene** common surfaces. If the surface has such a transition when exiting a volume, the corresponding frame on the other side of the common surface is inferred.

The algorithm assigns a property among `embedding, embedded, equal` or `disjoint` to all framed surfaces `i` on one side of a CS, comparing against all _relevant_ combinations of framed surfaces `j` on the other side. It uses the side division as an accelerating structure. Depending on the classification, transitions are assigned as following:
* `i` embedding `j`: `j` -> `i`
* `i` embedded in `j`: `i` -> `j`
* `i` equal to `j`: `i` <-> `j`
* `i` disjoint to all `j`: `i` -> -1 (no frame)
In the case of multiple embedding, the transition is assigned to the deepest frame.

The total percentage of frames having a pre-computed transition, and the percentage of transitions for common surfaces with frames on _both_ sides are reported. The performance gain compared to the previous version (one to the last column) seems to be highly dependent on these percentages.

![screen02](/uploads/8f69e701520cf816f70968db9a2c2d5d/screen02.png)

* Removed leftover comment

* Removed leftover debugging block

* Removed unneeded dependency

* Reverted to traversal index

* Multiple fixes for traversals

* Added iterator for slices

* Scene traversal working, started work on division-based acceleration

* still working...

* Divisions can return a list of candidates in an arbitrary range

* [surf_model] Implemented detection of disjoint frames.

The detection does not give false positives but gives false negatives. The algorithms must be improved to eliminate false negatives.

* No inline for device storage