use singleton for xercecs init and add destructor
This PR fixes massive memory leaks in the usage of xercesc.
The resources used by xercesc must be cleaned by calling `xercesc::XMLPlatformUtils::Terminate();`
This PR uses a singleton to ensure the correct initialization and termination of the data.

A destructor is added to clean up and a smartpointer is used for better memory management. 

Before this PR, it was hard to use valgrind because a huge amount of errors.

This cleans at least a big chunk of those errors. Calling 

```
valgrind --leak-check=full ./test/testRaytracing -gdml_name ../gdml/pcone_default.gdml  -verbosity 0 -debug 1 -only_surf 0
```
(on the surface model branch) yielded 1180 errors before, and 216 with this PR.

Cleaning to be continued...

* fix deleted comment

* use singleton for xerces init and add destructor