Change docker image location for CI
The VecGeom GitLab CI runs in a Docker VM image. It fails when pulling the image for the OS with 429 Too Many Requests due to Docker Hub plan limits, which seems to be now 10 pulls per IP address per hour.
This changes the image with a CERN pull-through cache for `ununtu-jammy`

* Change docker image location for CI