Enable single precision in solid BVH
This MR enables the compile option `VECGEOM_BVH_SINGLE` also for the solid model.

The BVH and AABB classes are templated for the precision type. The BVHManager and BVHNavigator are not, as the implementation should not touch the user interface.
Thus, the precision type Real_t is defined in both BVHManager and BVHNavigator depending on the compilation option.

Furthermore, a minor improvement is done: after the safety to out is calculated, it can be used to further reduce the limit that is passed to the safety to in calculation.

For locating, safety calculation, and raytracing, I get the following speedup for 10mio random rays in CMS2026D110 over using double precision (and no errors):

```
Locate: 7%
Safety: 22%
Traversal: 13%
``` 

An AdePT simulation with B field in FullSimLight got a 19% speedup from this MR.

Since we now do an approach and then calculate the distance, the difference in the distance can be larger than epsilon. Therefore the CI test is now adjusted that it uses kTolerance instead of epsilon as a criteria for near_equal.

* use kTolerance instead of epsilon

* fix CheckDaughterIntersectionsBenchmark

* update CMakeLists

* fix testraytracing after template removal

* cleaning

* revert templation approach to not touch user interfaces

* minor corrections

* enable solid BVH in single precision