Fixing the check for reference in TestBVHNavigator.
The raytracing of BVH and looper navigators are not numerically identical due to different optimizations, so this relaxes the criteria for valisation to the relative boundary error.

* Fixing the check for reference in TestBVHNavigator.

The raytracing of BVH and looper navigators are not numerically identical due to different optimizations, so this relaxes the criteria for valisation to the relative boundary error.

Fixed kRelTolerance when checking non-positive values.