Migrate to using CudaRdcUtils and fold vecgeom into vecgeomcuda
    VecGeom::vecgeomcuda and VecGeom::vecgeom has circular dependencies.
    vecgeomcuda uses the shape function implemented in vecgeom.
    vecgeom is requires some code that exist only in vecgeomcuda (eg.
    size of the some of the content of the namespace vecgeom::cuda)
    
    Users will link against VecGeom::vecgeom in all cases, in particular
    when building without CUDA enabled.
    
    If vecgeomcuda depends on vecgeom (the case for v1.2.10 and older)
    then users needs to explicitly add the dependency on vecgeomcuda
    when enable CUDA support.
    
    If vecgeom depends on vecgeomcuda, users don't need to change
    anything when enable CUDA support (and CudaRdcUtils will handle
    the dependency).
    
    However, if vecgeom depends on vecgeomcuda there is cases where
    some symbols in vecgeomcuda are not resolved (somehow the link
    should do another pass and/or we would need to add vecgeom 'again'
    after vecgeomcuda).
    
    Conclusion there is no good way to handle this split of library
    with circular dependency and as is, it does not help to build
    a non-cuda executable with a CUDA enable build of VecGeom.
    
    So there is 2 solution:
    1. Tweak the code and move things around to that vecgeom no longer
    depends on vecgeomcuda.
    2. Merge the 2 libraries into one.
    
    Until there is a strong business case for implementing 1., this
    commit does the much easier 2.

* Add backward compatible ALIAS VecGeom::vecgeomcuda_static

* Increment CudaRdcUtils version to 10

* Always explictly add cuda runtime to executable.

If the executable source uses the CUDA API only in C++ files,
cmake can not detect the need to add the CUDA runtime library.
However since the type of the runtime library (static vs shared)
depends on the time of the depend library, it is not easy for the
user to known which one to add.

For now, just always add the proper runtime library to the
executable runtime library.

* Detect late addition of CUDA source files.

If target_sources is called explicitly the target is turned into
an RDC executable only at the next call and to
cuda_rdc_target_link_libraries and there we need to make sure to
turn on the RDC linking.

Note: if the same happens for a library, it might not yet be
supported either.

* Increment CudaRdcUtils version to 8

* cuda_rdc_check_cuda_runtime: ignore non-targets.

An example of entity that will be a non target but is on the
list to be looked at include the target that were brought in by
inside a nested include and was not declared global; the target
introduced by that find_package will go "out of scope" when the
scope the `find_package` was called in ends.

* Prepend reversed order of (static) libs as needed for static builds.

* Change default to shared library ON

* fix typo in test/VecGeomTest/CMakeLists.txt.

This affects the case of G4 build without Geant4::G4persistency

* Fold libvecgeomcuda into libvecgeom

VecGeom::vecgeomcuda and VecGeom::vecgeom has circular dependencies.
vecgeomcuda uses the shape function implemented in vecgeom.
vecgeom is requires some code that exist only in vecgeomcuda (eg.
size of the some of the content of the namespace vecgeom::cuda)

Users will link against VecGeom::vecgeom in all cases, in particular
when building without CUDA enabled.

If vecgeomcuda depends on vecgeom (the case for v1.2.10 and older)
then users needs to explicitly add the dependency on vecgeomcuda
when enable CUDA support.

If vecgeom depends on vecgeomcuda, users don't need to change
anything when enable CUDA support (and CudaRdcUtils will handle
the dependency).

However, if vecgeom depends on vecgeomcuda there is cases where
some symbols in vecgeomcuda are not resolved (somehow the link
should do another pass and/or we would need to add vecgeom 'again'
after vecgeomcuda).

Conclusion there is no good way to handle this split of library
with circular dependency and as is, it does not help to build
a non-cuda executable with a CUDA enable build of VecGeom.

So there is 2 solution:
1. Tweak the code and move things around to that vecgeom no longer
depends on vecgeomcuda.
2. Merge the 2 libraries into one.

Until there is a strong business case for implementing 1., this
commit does the much easier 2.

* Remove explict dependence on vecgeomcuda (now implicit through vecgeom)

* Have libvecgeom depends on libvecgeomcuda.

This is the more natural dependency as user we will explicitly mention libvecgeom

* Import CudaRdcUtils v6: object lib and CUDA runtime.

This version avoids duplicating the object libraries onto unintended targets.
It also improves the handling of CUDA_RUNTIME_LIBRARY:
  - fix determination when not explicitly set
  - export CUDA_RUNTIME_LIBRARY to avoid guessing
  - fix the promotion and checks if there is a mismatch between current and dependent library

* Import CudaRdcUtils v4: improvement of linked library handling

* Add message about ignoring RDC installation

* Load CudaRdcUtils when loading VecGeom Targets

* Move CudaRdcUtils to cmake/external.

The official repository is https://github.com/celeritas-project/CudaRdcUtils

* CudaRdcUtils is changed, so update to v2

* CudaRdcUtils is changed, so update to v2

* Migrate to using CudaRdcUtils

Because we need to iterate through the list of sources and CMake
does not support nested list, we can not use generator expression
in the list of sources (nor in the list libraries)

* fix typo in casing of ROOT::Core

* fix silly copy/paste typo

* Add support for CMake 3.22 (only if no CUDA)

* fix typo

* Migrate to using CudaRdcUtils

Because we need to iterate through the list of sources and CMake
does not support nested list, we can not use generator expression
in the list of sources (nor in the list libraries)