Use exceptions for runtime errors, support assertions
Replaces !1304. `-DNDEBUG` is always used in vecgeom in practice, so all `assert` checks for runtime errors (e.g. out of memory) are compiled out. This adds two new macros: a `VECGEOM_ASSERT` for actual assertions, which throw `std::logic_error` on CPU and use `assert` on GPU; and `VECGEOM_VALIDATE` for runtime checks, which throw a `std::runtime_error` on CPU and print an error message (and abort the kernel) on GPU.

* Add on-device validate with message

* Don't replace RUNTIME_THROW with 'unreachable'

* Fix docs

* Fix runtime error message

* Fix validate -> assert

* Clang-format

* More adaptation to the portability macros.

* Fix macros

* Genericise brep manager

* fixup! Replace assert with VECGEOM_ASSERT

* Replace assert with validate

* Replace assert with VECGEOM_ASSERT

* Replace CUDA API calls and interface functions with assert macro

* Use API call macros for brep

* Define runtime macro

* Add vecgeom_assert

* Add multi-backend macros and functions